requirements.txt
## The following requirements were added by pip freeze:
aiofiles==23.1.0
async-timeout==4.0.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
hiredis==2.2.2
html5tagger==1.3.0
httptools==0.5.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
python-dotenv==1.0.0
redis==4.5.4
requests==2.29.0
sanic==23.3.0
sanic-routing==22.8.0
tomli==2.0.1
tracerite==1.1.0
ujson==5.7.0
urllib3==1.26.15
uvloop==0.17.0
websockets==11.0.2
